added change qty function in cart, update cart will successfully update the cart table accordingly.

+++ main_table.sql
@@ -3,7 +3,7 @@ use f37ee;
 create table products (
     id int unsigned not null auto_increment primary key,
     product_name varchar(255) not null,
-    price float(5,2),
+    price float(7,2),
     category varchar(255),
     description text(1000)
 );
